feat(order): add aut and routes

+++ src/main/java/com/tunde/GKwebhook/Public/domain/order/dto/InstalmentsDTO.java
@@ -2,6 +2,6 @@ package com.tunde.GKwebhook.Public.domain.order.dto;
 
 public record InstalmentsDTO(
         int numero_parcelas,
-        int valor_parcela
+        Double valor_parcela
 ) {
 }
